fixed a bit, random noise still does not work

+++ image.h
@@ -7,7 +7,7 @@ typedef struct Image {
     int height;   // Image height in px
     int channels; // Number of color channels (e.g. 1 = grayscale, 3 = RGB)
     int maxval;   // Maximum value of pixel value
-    double *data; // Pointer to image data, flat array of RGB pixels
+    int *data; // Pointer to image data, flat array of RGB pixels
 } Image;
 
 #endif
